Extension template machinery is working, I think,
Only the general_nonlinear_elasticity tmeplate is complete,
but it creates a Property that can be imported and used.

The old swig_sources function that used to be in the top oof2
CMakeLists.txt file had been moved to oofbuildtools.cmake so
that it could be used by extensions.  Some of its functionality
should have stayed in the main OOF2 file, so the top CMakeLists.txt
now defines a wrapper for it, called oof_swig_sources, and all
of the CMakeLists.txts in the subdirectories call that version
instead.