Improvements to the extension machinery.
Hid the OOF2COMMON and OOF2ENGINE variables in the extension's CMakeLists.txt.

Rewrote most of the extension README.

Minor cleanup in CMakeLists_subdir.txt.

Moved ramberg-osgood out of EXTENSION_TEMPLATES because it's not
a template.  It still needs to be updated.

Added OOF.File.Load.Module to make it easier to import a module,
and to log its importation.

Added a call to os.path.expanduser in the handling of the --pathdir
option.  Otherwise "oof2 --pathdir=~/mymodules" doesn't work.