Better error checking for non-square matrices, and a TODO.
Not all badly defined problems are caught by
SubProblemContext.checkSolvability().  The TODO in subproblemcontext.py
mentions one way of creating an uncaught problem.  Our wrappers to
the Eigen solvers now check that the matrix is square, which prevents
the program from crashing but doesn't provide much useful information
to the user.