Fixed various problems with the Ramberg-Osgood extension.
Updated the example.log example, which was still python2 and using
very old OOF menu items.  I don't think it's giving correct answers,
but the extension compiles and runs, which is what I was aiming for.

I deleted the README file, because it was about external extensions,
which this isn't anymore.