Added the start up option --separate-dialogs.
This turns off the transient_for option in gtk dialog boxes.
Transient dialogs are subwindows of their parent, which is
generally good, since they don't get lost that way.  But
the Mac screen capture utility can't automatically grab
a transient dialog, so it's easier when generating figures
for the manual to have dialog boxes be full fledged windows.