Moved Field::time_derivative() to CompoundField::time_derivative().
Putting it in Field was a hack to make FloatBCApp::editmap() work
on out-of-plane fields, which it needs to do when enforcing
periodic boundary conditions.  But only the CompoundField class
actually has auxilary time derivative fields.  Now FloatBCApp
has two versions of editmap, and the correct one is called by
FloatBCBase.applyBC.  This is not pretty but is less ugly.