Changed conjugate.conjugatePair so that it can use iterators instead of lists.
This makes it easier to use, since it can directly use Field.compoenents(),
etc.   There were two subtleties, however.

The FieldIndexes stored in the C++ ConjugatePair objects need to
be cloned, because the ones passed in were stored in deleted
IndexP objects.  So the clone() method had to be restored in FieldIndex.

Also, the iteration order matters.  The plane stress equation has
to match out-of-plane components of stress to displacement z-derivatives.
The order of the two is different -- the out-of-plane stress
components, as returned by Components, are zz, yz, xz, but
the displacement components  area x, y, z.