PropertyRegistration.fluxInfo and eqnInfo now check that they're being called with FluxProperties and EqnProperties respectively.
When registering a FluxProperty, PropertyRegistration.fluxInfo()
needs to be called to add Property metadata.  When registering an
EqnProperty, PropertyRegistration.eqnInfo() needs to be called to
add metadata.  Now those methods check that the Property is the
right type.  This is an abuse of good OO practice, but it's a bit
difficult to do better.  Maybe have two different kinds of
PropertyRegistrations?  That's ugly too.

To get this to work, swig needs to know whether Properties are
derived from FluxProperty or EqnProperty.  The swig files used to
ignore the intermediate classes and pretend that all Property
subclasses were derived directly from Property.  I swigged
AuxiliaryProperty too, just for completeness, although AuxiliaryProperties
don't have to add metadata.