All PropertyRegistration arguments are now key word arguments.
This should help prevent registration errors.  Also, in the previous
version propertyType was an optional argument, since it had a
default value, but was actually required.  I think it only had
a default argument only because it was after params in the argument
list, and params has a default argument.  I didn't want to document
that.  Now all args have default values, but assert statements
check that the required ones are set explicitly.
Please enter the commit message for your changes. Lines starting