FEMesh now stores a pointer to its swigged self.
When an FEMesh* is returned to Python, a swig typemap returns the
previously swigged object, so that data that was added in Python
is preserved.  This allows set_mesh_data() and get_mesh_data() to
work in Python Properties.

If a FEMesh is passed to Python wwithout going through SWIG, the
proper pointer is obtained from FEMesh::getPyMesh.  This happens
in the PyProperty wrapper functions that use the Python API to
call Python functions.