AuxiliaryProperties are now available in Python.
Classes can be derived from PyAuxProperty in the same way they can
be derived from PyFluxProperty and PyEqnProperty.  They're only
useful to other Python Properties, because other than cross_reference()
and other bookkeeping calls, none of a PyAuxProperty's methods are
available in C++.

Changed pyheatconductivity.py and pyproperty_test.py so that the
PyAuxProperty is tested, along with precompute() and set_mesh_data()
in PyFluxProperty.