Did a longstanding TODO about a better way to evaluate Fields inside Property methods.
The old way used the generic OutputValue machinery, which was
cumbersome and expensive (and I didn't want to include it in the
manual).  Now it uses Field::value() and Field::gradient(), which
are much easier to call.  Under the hood the base class methods
still use the OutputValue machinery, but they can be redefined in
derived classes.  That has been done for ScalarFields, which covers
all instances in the current set of Properties.

Added Element::interpolate_deriv(), which is the same as
Element::interpolate(), but uses the shape function derivatives.