Fixed the integration_order() methods in Properties.
Most but not all were technically wrong, but probably had no
consequences except in higher order elements than the ones we have.
The polynomial degree of the test function that multiplies the
integrand (and is integrated by parts in divergence terms)
wasn't being included.

None of the regression tests were affected by this change.