Minor changes.
Added a missing DECREF in typemaps.swg, and a comment saying that it's
probably inconsequential, but technically correct.  The FETCH_CLASS
swig macro keeps a reference to a class, but classes are not likely
to be destroyed.

Added a repr for EmptyTip in parameter.py.  This made some debugging
output easier to read.