Simplified functions in cstrain.C.  Fields can return all components in a single call.
Added Field::values and Field::gradients, which are non-virtual
methods that return all components of a Field or its gradient in
a single call.  The type of the returned object depends on the Field
subclass.  This avoids the clunkiness of the generic Output mechanism,
which works on a generic Field pointer.  The new methods are easier
to use but require that a Field pointer be cast to its derived class
first.

The functions in cstrain.C now use the new Field methods.  They also
return vector or matrix objects instead of requiring a target object
to be passed as an argument.  This is okay because the vector and
matrix classes have move constructors, so no extra copies are made.