Added a localdata argument to FluxProperty methods.
This makes begin_point() and end_point() much more useful.
begin_point() can allocate and return a pointer to an object,
which will be passed to all of the other methods (flux_matrix(),
etc) during the linear system construction process.  end_point()
should delete the object.   The object is passed as a data* in C++,
and just a PyObject* in Python.

I did this now because when documenting begin_point() and end_point()
it seemed silly to tell people that the other methods could use
data created by begin_point(), but not give any mechanism for doing
so.

pyelasticity.py includes a simple check that data created by
begin_point() in Python is correctly passed through C++ to
the other Python methods.