Changed the sign of the Stress flux, back to the more natural definition.
An earlier version of Eigen didn't work with the force balance
equation in the form we had been constructing it, because apparently
the stiffness matrix was negative definite.  (The CG solver we had
been using before switching to Eigen was happy to solve the negative
definite matrix).  To fix that problem without breaking any user
code we had changed the sign of the Stress internally, and also
added a minus sign to the output.  That problem doesn't seem to
exist with the current Eigen, so the signs have been changed back.
This makes it easier to document the FluxProperty methods, because
no weird minus signs have to be explained.

HOWEVER, I had to increase the max_iterations on many tests
after these changes.  I don't understand that.  The modified
tests will be in the next commit.

Added some TODOs about eliminating loops by using vector and
tensor operators in the Property methods.  This would make the
code neater but probably wouldn't provide any speedup.