Increased max_iterations and/or file comparison tolerances for Stress tests.
Recent sign changes in the force balance equation (see previous commit)
required some changes in the tests.  Some were understandable -- the
numerical noise was different so either the reference files needed to
be changed (next commit) or tolerances increased.  But some tests
failed unless max_iterations for the matrix solver was increased.
I don't understand this and will investigate it next.

In solver_test.py, htere is now a global variable for setting
max_iterations, and an easier way of running a small number of
tests for debugging.