Changed the preconditioner in basic solver mode back to ILUT.
Also added more comments about the preconditioner issue.

The dilemma is this.  The way we defined the force balance equation
creates a negative definite stiffness matrix.  Conjugate gradient
requires a positive definite matrix.  In the distant past, that was
somehow ok -- the solvers we were using could handle it.  But we
started using Eigen, and we had to fix the problem, which we did
by effectively changing the sign of the stress, and adding a minus
sign to all stress outputs.  That was ugly but worked.  However, I
don't want to have to document the extra minus sign in the manual,
because there ought to be a better way of handling it.  So I took
it out, and discovered that the current Eigen (3.4.0) *does* work
with the negative definite matrices, apparently.

In a separate but related set of commits, we changed the default
preconditioners.  I think, but am not absolultely kure, that ILU
should not work with positive definite matrices and CG, because the
preconditioned matrix doesn't have the right symmetry.  IC preserves
the symmetry, so we made it the default for everything.  However,
after changing the sign of the force balance equation, IC started
requiring a lot more iterations than it used to.  Changing the sign
of the matrix and rhs by mulitplying them by -1 just before calling
the matrix solver (see subproblemcontext.py in this commit) brought
the iteration count under control, presumably meaning that IC
requires a positive definite matrix.  Hacking in a minus sign isn't
a good general solution, since it won't necessarily work for coupled
problems.

However, CG with ILU *does* work, and quickly, with the current
Eigen, although I don't know if it's supposed to.  So this commit
changes the default preconditioner back to ILU.  It passes all the
tests, so CG+ILU is getting the right answers.  Maybe Eigen has
implemented ILU in a way that preserves positive definiteness?

The next commit updates the test, making them use ILU too.

See these commits:

Changed the sign of the force balance equation
9b8bce21091b829c74359d1b954da6365855f673

Changed the sign of the Stress flux, back to the more natural definition
e034712734192ac6a4eba93841feab590b6a560f

Added a "negate" flag to the Flux class
d340d76d0703108c80741fbfbe2c7a2f0056bd97

Removing the extra minus sign
cd8c886c5b3bb2b2783bda5ce703269991221968
and subsequent commits

01bbe20ed8e029fc935f73b17fe179a5f43deb42
Added an Incomplete Cholesky preconditioner, and used it in all tests that used to use ILU with CG.

Also see "git log --grep ILU".