Copied minor changes from undo-elasticity-sign-change, which is being abandoned (maybe).
Marked points that need attention with TODO_FLUX_SIGN.  These may just
need comments.

I now think that the problems with the flux signs that I discovered
while working on the remove_static_flux_value branch and was trying
to fix with the undo-elasticity-sign-change branch were due to
inconsistent notation, which was foiling our attempts to make
the flux and equation handling generic.

The stress flux is defined as modulus * field gradient, and
the force balance equation is  mass*acceleration - div stress = f.
However, the heat flux is defined as _negative_ modulus * field gradient,
and the heat equation is C*dT/dt + div flux = heat source.