Changed two menu item names for consistency.
Add_property and Remove_property were the only menu items that
used a lowercase letter after an underscore.  Now they're
Add_Property and Remove_Property.  The old names are still
in the menus, but they're clones of the new menu items and are marked "secret".

Added a "secret" kwarg to OOFMenuItem.clone().