Secret menu items don't appear in the manual.
The manual pages for the secret menu items that preserved backwards
compatibility when Add_property and Remove_property were changed
to Add_Property and Remove_Property were hiding the pages for the
new menu items.  Probably because of case insensitivity in the links
and/or the file names (on Mac).