Deleted the IndexTypes classes and replaced them with ints.
The classes were used sort of arbitrarily.  There were a
lot of cases in which it wasn't clear if a variable should
have been a SpaceIndex or not.  Since I didn't want to
have to document it, and because the benefit provided by
the classes was minimal, I just deleted them.

Element::outputFieldDeriv and Element::outputFieldDeriv had
SpaceIndex* arguments, so that a typemap could be used to convert
an Enum to a SpaceIndex.  Now the args are just ints, not pointers.