Improvements to DoubleVec.
It now supports the usual arithmetic operations, both in C++
and Python.  The Python versions do all their looping in C++,
so there shouldn't be much of a performance penalty.

Removed DoubleVec::scale() because it's now redundant with
DoubleVec::operator*=() in C++ or DoubleVec.__imul__ in Python.

Deleted the DoubleVecIterator class.  DoubleVec::iterator is
now an Eigen::VectorXd::iterator.

DoubleVec::__getitem__ and DoubleVec::__setitem__ are left in,
because they're use by the test suite.  They shouldn't be used in
production code.  There no way of iterating over a DoubleVec in
Python, except by using __getitem__ and __setitem__ in a loop over
the index.  This also should not be done in production code.

Added tests for most DoubleVec methods.