Improved debugging tools.
getrefcount(obj) returns the number of references there are to
the python object obj.  It's just a wrapper around sys.getrefcount()
that takes into account the extra references from the additional
layer of the call stack.

Added callerName(), which returns a string that is simpler than
the string that dumpCaller() prints.

Improved the comments.