Debugging the DoubleVec destruction problem.
The problem is in the line "values += s*update" in Newton.solve in
nonlinearsolvercore.py.  I don't actually know what's wrong with
it, but I have a guess.

I'm committing this now to save the debugging lines.  They have
mostly been commented out, and will probably be deleted in a later
commit.