Fixed DoubleVec, I think.
The in-place arithmetic operators (operator+=, etc) have to return
void.  If the swigged function returns *self, then python thinks
that a reference is being discarded, because the return value
isn't being assigned to anything.

The test case where this problem arose is still not working.  It
looks like it might be in an infinite loop.  But it's getting
farther than it did before, and not seg faulting.