Added a new simpler version of ScriptLoader, which sort of solves some problems, maybe.
What it *doesn't* do is allow a progress bar to be displayed.

What it is an attempt to do is to run functions defined in
the script correctly.  The old code didn't set globals correctly
inside the function, so local variables defined in the calling
scope weren't available.  I'm not sure that the new code
actually solves the problem, though.

Some of the GUI error handling tests were failing, becaue the new
ScriptLoader produces different error messages.  Now they can work
with both versions of ScriptLoader, via some hacking which should
be removed when ScriptLoader is really fixed.