Working on modifying ScriptLoader.
The old version didn't put the correct variables in global scope
when evaluating functions defined in the script.

Passing globals|locals as the "globals" arg to eval() seems to
fix the immediate problem.

The current problem is in passing exceptions out from the script,
and presenting useful information to the user.  Something changed
in the old code for doing that and has broken the tests. The old
code was a mess anyway.  I'm currently trying to fix that, possibly
using Exception.__cause__, which track exceptions raise while handing
other exceptions.

Committing on "inprogress" branch, to save work in progress,
including lots of debugging lines.

A lot of the tests were modified because the error messages
changed.  This is a pain because the error format also
changes with different Python versions and the tests need
to work with multiple versions (3.9 to 3.13 at the moment).