Working on debugging an issue that only appears on the Mac, as far as I can tell.
I am going to temporarily abandon this issue and this branch,
so this commit includes a bunch of half-baked ideas.

To exhibit the bug, run oof2 --debug --script=bigms.py,
where bigms.py contains these lines:
X-----
w = 100
h = 800

OOF.Microstructure.New(name='microstructure', width=1, height=1, width_in_pixels=w, height_in_pixels=h)
OOF.File.Save.Microstructure(filename='junk.ms', mode='w', format='ascii', microstructure='microstructure')
X----

and then save the microstructure again, using the File/Save/Microstructure menu item in the GUI.

Here are the comments from SCRATCH/bigms.py, which explains the bug:
XX---------
Load this file and save the Microstructure with the GUI.

This crashes on the Macs.  Using big-ms-bug branch.

It doesn't crash if run with --unthreaded or in text mode.  Is this
is a race condition?  The failure is very repeatable, making a race
condition unlikely.

It doesn't crash if the Save command is in the script, but it does
crash if the Save command is run again after the script ends.

It doesn't crash if the sections of writeMicrostructure()
microstructureIO.py that call getCategoryMapRO() and use its results
are commented out.

It doesn't crash if the GUI is open but the "OOF.File.Save..."
command is entered in the Console Window.

It doesn't crash on Linux.

It doesn't crash if the size in pixels is smaller.
  w   h
 400 300   no crash
 500 400   no crash
 600 500   no crash
 800 500   no crash
1000 500   no crash
 100 600   crash
 200 600   crash
 700 600   crash
 800 600   crash
1000 600   crash

No, it doesn't crash if h is smaller than 550
 100 500   no crash
 100 525   no crash
 100 538   no crash
 100 544   no crash
 100 547   no crash
 100 548   no crash
 100 549   no crash
 100 550   crash    <-- can't reproduce this
 100 600   crash

No, it *sometimes* doesn't crash if h is smaller, but can crash if
the Save is repeated. Smaller h allows more iterations before
crashing.
 100 100   no crash in 10 iterations
 100 300   crash on iter 3 (3 times)
 100 500   crash on iter 2 (3 times)
 100 549   crash on iter 2 (3 times)
 100 550   crash on iter 2 (6 times)
 100 575   crash on iter 2 (6 times)
 100 600   crash on iter 1 (6 times)

  10 571   crash on iter 2

The total number of pixels doesn't matter, just h
    1 10000 crash on iter 1
   10 1000  crash on iter 1
  100 100   no crash
 1000 10    no crash
10000 1     no crash

oof2 2.3.0 crashes, but not on the same iterations
 800 700    crash on iter 1
 800 600    crash on iter 2  (was iter 1 w/2.3.4)
 800 500    crash on iter 2
 100 500    crash on iter 2
XX-----