Improvements to error handling and reporting, motivated by changes in ScriptLoader.
If an ErrError exception is converted to another type of exception,
the error reported is now the original exception, not the secondary
one.  (The secondary one could be a ScriptException, if the problem
occurred while loading a script.)

Fixed line spacing in the error pop-up window.

ScriptInterrupt and ScriptException are now sublasses of PyOOFError,
not just Exception.

Added ErrError::operator==()

Added a way for test scripts to find out the location of C++ source
files, so that ErrProgrammingErrors that incorporate the file
name in the message can be tested.  This is currently in ooferror.C,
but should probably be moved, maybe to cdebug.C.

All of the PyOOFError classes that are created from the ErrError
classes by the WRAP_OOFERROR_CLASS swig macro are now injected
into the main oof2 namespace.  I'm not sure this is absolutely
necessary, but seems like a good idea.

Fixed some of the tests in fundamental_test.py that were broken
by the ScriptLoader changes.