May have fixed the DoubleVec problem.
Will need to remove debugging lines and run the full tests to
be sure(r).

The problem was swig was generating bad code for
DoubleVec::operator+=() and other in-place operators.  The return
value of the operator was a new Python wrapper for the underlying
C++ object, so the destructor was being called too many times
when the wrappers went out of scope.

Also, even though it's not needed in C++, the operators must
return a reference to themselves.