Removed the assertion about the order of the bounds in LogDefiniteProgress.setRange().
And added a comment.  If the bounds are out of order, presumably
there will be no iterations of whatever procedure is displaying the
progress bar.  So it doesn't matter if what the range is, because
the progress bar will never be updated. Putting the assert statement
there was making tests fail when run in debug mode.