Removed the two argument version of DoFMap::inject(), which was being used incorrectly.
In substance, the two and three argument versions differed only in
that the two argument version assumed that the offset was zero.
But the two versions had different assert statements, which were
incorrect in one version. (It didn't account for extra time derivative
data appended to the dof vectors.) So when compiled in debug mode,
the program failed some tests which it shouldn't have failed.

I also removed inject() from the swig file, since it's not used
from C++.