Added MessageManager.latest(), to allow regression tests to see what's been printed.
It was previously only possible to test printed output in GUI tests,
where the message window's contents could be inspected.  Non-gui
methods that wrote to stdout couldn't be tested.  MessageManager.latest()
returns the last n lines printed by the various methods in reporter.py,
so it's not actually testing what's going to stdout or stderr, but
it's close.