Fixed a problem in fundamental_test.py.
Error messages from C++ often include the source file path. The
path depends on where the program was build, the the machinery has
to take that into account when checking that error messages are
correct.  It was incorrectly assuming that the source directory was
named OOF.

The new version uses a CMake variable to set oofSourceDir in oof2-test,
instead of the previous ugliness.