Trying to get the old numpy branch to build.
This is the branch with the original numpy code, before merging
with the develop branch.  That merge is on switch-to-numpy-merge.
The merged code crashes when saving an image with scikit-image,
and I want to see if the original branch has the same problem.
So I'm doing what I can to get the original branch to build,
without copying any numpy or scikit-image code from the merge.

I'm not entirely sure why the original branch didn't build.  It's
hitting a swig syntax error.  I'm not sure it ever built in the
form it was in, or maybe there was a swig change.  I don't want to
install an old swig right now.