Did a minor TODO.  Also renamed image test reference files.
When not using ImageMagick, adding OOFCANVAS_CFLAGS to all header
files is not necessary, as suspected.

Image test reference files were renamed so that they can easily
be compared to test results outside of oof2.  They need to have
a file suffix that reflects the image type.

I didn't realize that "git commit" would also commit the results of
"git mv" when the moved files hadn't been added.  Those don't belong
in the same commit as the TODO but I'm not going to undo it.