Added OOFImage::enhance_contrast().
The scikit-image enhance_contrast function requires the image
channels to be 8 bit unsigned chars.  OOFImage::enhance_contrast(),
in constrast.C, works with doubles and can be easily converted
to other numerical data types.

C++ methods that use the numpy data stored in an OOFImage object
now expect it to be a PyArrayObject*, not a PyObject*.