Various image modification modifications.
Moved the denoising modifiers from imagemodifier.py to
denoise.py, because they were taking too much space.

Attempted and failed to reproduce the ImageMagick reduce_noise
method.  I don't believe that the method does what the documentation
says that it does.  The new code is in reducenoise.C, and the
Registration for it is in denoise.py, but is marked secret.

Added comments to contrast.C and removed old code.

Other fixes to comments and TODOs.