Gray scale images can now be loaded, manipulated, and saved without converting to RGB.
This was a long standing TODO that I had no intention of every doing,
but it solved a problem with the image modifiers, and actually
(knock wood) wasn't too difficult to implement.  The immediate
problem that it solved is that the AddNoise modifier doesn't add
RBG noise to a gray image.