More fixes indirectly related to numpy/skimage related changes.
Error messages produced by gui tests weren't printing source file
paths correctly.  The source directory used to just be set in
oof2-test, using a cmake substitution in oof2-test.in.
Now it's also set in oof2.in, and it's stored in an environment
variable so that it works no matter how oof2 is started.

Two test files were updated to use the new way of fetching the
source directory.  There are probably others to fix as well.

Also added better error messages in errorMsg(...) in generics.py.