Fixes for image handling.
File types are checked. It's an error to ask for a file type that's
not supported by PIL.

Tiff files are handled separately.  They can be saved without crashing,
but aren't necessarily readable on all systems.  I don't know what's going
on there.

Fixed a bug in BlurImage.  It was setting channel_axis incorrectly
for grayscale images.