Fixed image saving, I think, and updated tests for it.
Some of the various grayscale image formats have to saved by different
routines.  For most it seems to be impossible to preserve a meaningful
bit depth and to avoid converting to RGB or RGBA.  The numpy (npy
and npz) and tiff formats preserve all data.  png is written in 16
bit gray, so it doesn't lose much data.  All other types quantize
the data at 8 bits and/or save redundant RGB or RGBA channels.

Refactored oofimage.readNumpyImage().  The part that reads an
image file and creates numpy data is now done by getNumpyData(),
which is used in the OOF_Image("SaveGray") test.

Many more formats are tested explicitly in OOF_Image("SaveGray").